pyproject.toml
[tool.poetry]
name = "support-chat"
version = "0.1.0"
description = ""
authors = ["Jayesh Patil <jayeshvpatil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.5"
streamlit = "^1.31.0"
chromadb = "^0.4.22"
vertexai = "^0.0.1"
google-cloud-aiplatform = "^1.41.0"
pypdf = "^4.0.1"
sentence-transformers = "^2.3.1"
docarray = "^0.40.0"
google-api-python-client = "2.95.0"
google-api-core = "2.11.1"
html2text = "^2020.1.16"
faiss-cpu = "^1.7.4"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
